backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.35
alembic==1.13.3

# Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment Variables
pydantic-settings==2.1.0
python-dotenv==1.0.0

# File Upload
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development Tools (Optional)
black==23.12.0
flake8==6.1.0
